Started updating parser
Changed the error handling in the parser to be via the compiler instead of via tokens.
The error token is now only used to report a lexing error
This will allow for slightly prettier error messages and make other things easier to add (e.g. part of the dependency checking to make c-like syntaxes possible with the out-of-order compilation without using lots of keywords)